Edit header
Header was messy, navigation links overlapped each other and it was
hard to navigate around the webpage.

Each of the header's element is now placed so it has space around
and it is easy to navigate. Besides, header is responsive now and
changes depending on how much space has available.